+++ contrib/bindings/python/pathrs/pathrs_build.py
@@ -58,9 +58,9 @@ def create_ffibuilder(**kwargs):
 
 	return ffibuilder
 
-def find_ffibuilder():
+def find_rootdir():
 	# Figure out where the libpathrs source dir is.
-	ROOT_DIR = None
+	root_dir = None
 	candidate = os.path.dirname(sys.path[0] or os.getcwd())
 	while candidate != "/":
 		try:
@@ -69,36 +69,65 @@ def find_ffibuilder():
 			with open(candidate_toml, "r") as f:
 				content = f.read()
 			if re.findall(r'^name = "pathrs"$', content, re.MULTILINE):
-				ROOT_DIR = candidate
+				root_dir = candidate
 				break
 		except:
 			pass
 		candidate = os.path.dirname(candidate)
 
-	# TODO: Support using the system paths.
-	if not ROOT_DIR:
-		raise RuntimeError("Could not find pathrs source-dir root.")
+	if not root_dir:
+		raise FileNotFoundError("Could not find pathrs source-dir root.")
+
+	return root_dir
+
+def srcdir_ffibuilder(root_dir=None):
+	"""
+	Build the CFFI bindings using the provided root_dir as the root of a
+	pathrs source tree which has compiled cdylibs ready in target/*.
+	"""
+
+	if root_dir is None:
+		root_dir = find_rootdir()
 
 	# Figure out which libs are usable.
-	lib_paths = []
-	for mode in ["debug", "release"]:
-		so_path = os.path.join(ROOT_DIR, "target/%s/libpathrs.so" % (mode,))
-		if os.path.exists(so_path):
-			lib_paths.append(so_path)
-	lib_paths = sorted(lib_paths, key=lambda path: -os.path.getmtime(path))
-	lib_paths = [os.path.dirname(path) for path in lib_paths]
+	library_dirs = (
+		os.path.join(root_dir, "target/%s/libpathrs.so" % (mode,))
+			for mode in ("debug", "release")
+	)
+	library_dirs = (so_path for so_path in library_dirs if os.path.exists(so_path))
+	library_dirs = sorted(library_dirs, key=lambda path: -os.path.getmtime(path))
+	library_dirs = [os.path.dirname(path) for path in library_dirs]
 
 	# Compile the libpathrs module.
-	return create_ffibuilder(include_dirs=[os.path.join(ROOT_DIR, "include")],
-	                         library_dirs=lib_paths)
+	return create_ffibuilder(include_dirs=[os.path.join(root_dir, "include")],
+	                         library_dirs=library_dirs)
 
-if __name__ == "__main__":
-	# Compile the cffi module if running outside of setuptools.
-	ffibuilder = find_ffibuilder()
-	ffibuilder.compile(verbose=True)
-else:
-	# Use the system libraries if running inside setuptools.
-	ffibuilder = create_ffibuilder(include_dirs=[
+def system_ffibuilder():
+	"""
+	Build the CFFI bindings using the installed libpathrs system libraries.
+	"""
+
+	return create_ffibuilder(include_dirs=[
 		"/usr/include",
 		"/usr/local/include"
 	])
+
+if __name__ == "__main__":
+	try:
+		# Search for the compiled libraries to link to from our libpathrs
+		# source if running outside of setuptools as a regular program.
+		ffibuilder = find_ffibuilder()
+	except FileNotFoundError:
+		# If we couldn't find a valid library in the source dir, just fallback
+		# to using the system libraries.
+		ffibuilder = system_ffibuilder()
+	ffibuilder.compile(verbose=True)
+elif os.environ.get("PATHRS_SRC_ROOT", "") != "":
+	# If we're running in setup tools, we can't easily find the source dir.
+	# However, distributions can set PATHRS_SRC_ROOT to the path of the
+	# libpathrs source directory to make it easier to build the python modules
+	# in the same %build script as the main library.
+	ffibuilder = srcdir_ffibuilder(root_dir=os.environ.get("PATHRS_SRC_ROOT"))
+else:
+	# Use the system libraries if running inside standard setuptools.
+	ffibuilder = system_ffibuilder()